+++ Coimbra/DisableResizeAttribute.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Coimbra
+{
+    /// <summary>
+    /// Apply to a <see cref="SerializableList{T}"/> or <see cref="SerializableDictionary{TKey,TValue}"/> to lock its size in the inspector.
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Field)]
+    public sealed class DisableResizeAttribute : Attribute { }
+}