+++ Coimbra/ISerializableCollection.cs
@@ -1,6 +1,0 @@
-﻿using UnityEngine;
-
-namespace Coimbra
-{
-    internal interface ISerializableCollection : ISerializationCallbackReceiver{}
-}