+++ Coimbra/LockDictionarySizeAttribute.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-namespace Coimbra
-{
-    /// <summary>
-    /// Apply to a <see cref="SerializableDictionary{TKey,TValue}"/> to lock its size in the inspector.
-    /// </summary>
-    public sealed class LockDictionarySizeAttribute : PropertyAttribute { }
-}