+++ Coimbra/TypeFilterAttribute.cs
@@ -6,6 +6,7 @@ namespace Coimbra
     /// <summary>
     /// Apply this to either a <see cref="ManagedField{T}"/>, or <see cref="Reference{T}"/>, or a field with <see cref="TypeDropdownAttribute"/> to filter which types should be available.
     /// </summary>
+    [AttributeUsage(AttributeTargets.Field)]
     public class TypeFilterAttribute : TypeFilterAttributeBase
     {
         /// <summary>