+++ Coimbra/TypeFilterAttributeBase.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using UnityEngine;
 using Object = UnityEngine.Object;
 
 namespace Coimbra
@@ -7,7 +6,8 @@ namespace Coimbra
     /// <summary>
     /// Implement this to filter which types should be available to either a <see cref="ManagedField{T}"/>, or <see cref="Reference{T}"/>, or a field with <see cref="TypeDropdownAttribute"/>.
     /// </summary>
-    public abstract class TypeFilterAttributeBase : PropertyAttribute
+    [AttributeUsage(AttributeTargets.Field)]
+    public abstract class TypeFilterAttributeBase : Attribute
     {
         /// <summary>
         /// Return true if the <paramref name="type"/> should be allowed, false otherwise.